tools/mock_server/requirements.txt
Django==1.10
Faker==0.7.3
djangorestframework==3.4.7
